requirements.txt
Flask
requests
pytesseract
pdf2image
pandas
gspread
oauth2client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

Pillow
